RABSW-653 Data Movement - Lustre CSI Driver (#2)
In lustre mode, add a call to os.MkdirAll() to create the mountpoint where
CSI wants it, deep down in /var/lib/kubelet/pods...  K8s will look for a
'volumeMounts' section in a Pod's definition to indicate where the FS should
be bind-mounted into the pod's namespace.

Simplify the config/ dir.  The default is mock, and works for KIND.  Add an
OVERLAY=lustre to enable the lustre bits, which works for craystack-lop.
Throw out everything else.

Provide a config/samples/standalone for exercising this without having
lustre-fs-operator managing the PVs and PVCs.  This works for KIND.

When in mock mode, add a log message to say so.