Switch to Go 1.19.
Go 1.18 requires a switch from 'go get' to 'go install' for the retrieval
of the kustomize tool.  This is taken from the makefile
boilerplate that is generated by the current kubebuilder tool.

Update the version of kustomize from 3.8.7 to 4.5.7, because the Go 1.19 'go
install' command does not find the earlier version when run on an M1.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>