Fixups for the release workflow tools. (#34)
On the master branch, allow git-version-gen to use the SHA as the value
for the .version file, because there are no tags on the master branch for it to
use for that purpose.  This allows nnf-deploy to succeed when using this master
branch in a KIND environment.

Rename the workflow that verifies tags.

Build and do tags on "releases/v*" branches.

Weed out the "UNKNOWN" from git-version-gen.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>
Signed-off-by: Matt Richerson <matthew.richerson@hpe.com>