Update the README and the helm chart (#95)
Update the supported version matrix in the README. Fix some lint issues, and
update the instructions to deploy and release.

Update the chart to add the release version as a label on the resources.

Update the namespace spec to add an annotation to prevent helm from deleting
the namespace during 'helm uninstall'.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>
Signed-off-by: Anthony Floeder <anthony.floeder@hpe.com>